+++ shiny/_versions.py
@@ -1,6 +1,6 @@
-shiny_html_deps = "1.11.1.9001"
+shiny_html_deps = "1.12.1.9000"
 bslib = "0.9.0.9002"
-htmltools = "0.5.8.9000"
+htmltools = "0.5.9.9000"
 bootstrap = "5.3.1"
 requirejs = "2.3.6"