+++ shiny/www/shared/prism-code-editor/languages/ini.js
@@ -1,0 +1,7 @@
+import { l as languageMap } from "../index-CKRNGLIi.js";
+languageMap.ini = {
+  comments: {
+    line: ";"
+  }
+};
+//# sourceMappingURL=ini.js.map